# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WatchStore Api List",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/v1/send_sms": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for send verification sms to user",
                "operationId": "d66c863c38e88d858c0ed44f85519d62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "Enter mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/verify_sms_code": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to check sms code that recieved by user",
                "operationId": "e5905d20c2f34f98c4a5c597eefbd455",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "user mobile number",
                                        "type": "string"
                                    },
                                    "code": {
                                        "description": "recieved user sms code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth Api",
            "description": "Auth Api"
        }
    ]
}